+++ src/engine/Algebra.ts
@@ -594,3 +594,14 @@ export class GlobalCoordinates {
 
   constructor(public worldPos: Vector, public pagePos: Vector, public screenPos: Vector) {}
 }
+
+/**
+ * Shorthand for creating new Vectors - returns a new Vector instance with the
+ * provided X and Y components.
+ *
+ * @param x  X component of the Vector
+ * @param y  Y component of the Vector
+ */
+export function vec(x: number, y: number): Vector {
+  return new Vector(x, y);
+}